requirements.txt
Pyrogram>=2.0.33,<2.1
TgCrypto==1.2.3

python-dotenv>=0.10
